mcpserver: consolidate tool test helpers into helpers_test.go
