+++ mcpserver/tools/agents_test.go
@@ -9,7 +9,6 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/mattermost/mattermost/server/public/model"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -30,12 +29,6 @@ func newTestAIBotsServer(t *testing.T, bots []AIBotInfo) *httptest.Server {
 	return httptest.NewServer(mux)
 }
 
-func newTestClient4(serverURL string) *model.Client4 {
-	client := model.NewAPIv4Client(serverURL)
-	client.AuthToken = "test-token"
-	return client
-}
-
 func TestListAgents(t *testing.T) {
 	sampleBots := []AIBotInfo{
 		{ID: "bot1id12345678901234567", DisplayName: "Otto", Username: "otto"},
@@ -47,7 +40,7 @@ func TestListAgents(t *testing.T) {
 
 	t.Run("marks self agent", func(t *testing.T) {
 		provider := newTestProvider(t, ts.URL)
-		mcpCtx := &MCPToolContext{BotUserID: "bot1id12345678901234567", Client: newTestClient4(ts.URL)}
+		mcpCtx := &MCPToolContext{BotUserID: "bot1id12345678901234567", Client: newTestClient(ts.URL)}
 		argsGetter := func(target any) error {
 			return json.Unmarshal([]byte(`{}`), target)
 		}
@@ -59,7 +52,7 @@ func TestListAgents(t *testing.T) {
 
 	t.Run("unreachable server", func(t *testing.T) {
 		provider := newTestProvider(t, "http://127.0.0.1:1")
-		mcpCtx := &MCPToolContext{Client: newTestClient4("http://127.0.0.1:1")}
+		mcpCtx := &MCPToolContext{Client: newTestClient("http://127.0.0.1:1")}
 		argsGetter := func(target any) error {
 			return json.Unmarshal([]byte(`{}`), target)
 		}
